𒀯𒀲𒁘𒁃𒁔𒁎𒀚𒀄𒀱𒀏𒀯𒁁𒁆𒁏𒁆𒀗𒀙𒀓𒀆𒀏𒀽𒀑𒀨𒁇𒀆𒀃𒀚𒁟𒀱𒀔𒁂𒀷𒀍𒀾𒀿𒁝𒀃𒀭𒀣𒀓𒀓𒁡𒀶𒀦𒀶𒁆𒀸𒁞𒀆𒀧𒀎𒀆𒀍𒀉𒀑𒀝𒀫𒀿𒀷𒁐𒀭𒁔𒁘𒁑𒀃𒁈𒁗𒀨𒁘𒀟𒀂𒀲𒀎𒁖𒀫𒁛𒀏
F40F657874726163742068616E646C65722066756E6374696F6E7320696E746F1700F00E732F206469726563746F727920286D782331393220505220342F313429

[keccak512:cuneiform|lz4:base16]
